requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
debugpy==1.8.14
dill==0.4.0
distro==1.9.0
faker==37.1.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
jiter==0.10.0
mccabe==0.7.0
mypy-extensions==1.1.0
numpy==2.2.5
openai==1.79.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pip==25.1.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-core==2.33.2
pydevd-pycharm==251.25410.122
pylint==3.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
tomlkit==0.13.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
